Format returned value to string
I was surprised the bot didn't notice a new article on the link I parse.
After a few hours of debugging, I noticed here I don't format the list, hence the code breaks;
but because that wasn't logged, I didn't see what's going on, didn't see the error why it couldn't see the new entry and post that.