Bugfix: parse the RSS feed periodically
I was hunting the bug that the bot doesn't "see" new entries on the feed for a month.
It was able to get and record the currently available feeds at start-up, but it couldn't see the new entries in the endless loop. Sure, because it didn't reach out to the RSS feed to see if there are new articles, it was comparing the already fetched list's first element.